requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.0
webdriver-manager==4.0.1
openai==1.3.0
pandas==2.1.3
openpyxl==3.1.2
xlsxwriter==3.1.9
numpy==1.25.2
scikit-learn==1.3.2
matplotlib==3.8.1
seaborn==0.13.0
nltk==3.8.1
textblob==0.17.1
wordcloud==1.9.2
python-dotenv==1.0.0
fake-useragent==1.4.0
streamlit==1.28.1
plotly==5.17.0